+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v0.2.2 (2024-01-04)
+
+### Added
+
+- `Size` and `Point` now implement both to/from winit's
+  PhysicalSize/PhysicalPosition for both `i32`/`Px` and `u32`/`UPx`.
+
 ## v0.2.1 (2021-12-28)
 
 ### Added